requirements.txt
psycopg2-binary
tqdm
pyinstaller
tabulate